shell

+++ common/malloc.c
@@ -146,7 +146,7 @@ void free(void* ptr)
 	if (next != brk && *next & 1) {
 		fin = next + size(next) ;
 		unlink(next) ; }
-	if (deb != end *prev & 1) {
+	if (deb != end && *prev & 1) {
 		if (*prev & 8) deb -= *prev >> 4 ;
 		else deb = **(uint8***) (prev-4) ;
 		unlink(deb) ;	}
